requirements.txt
Flask
rembg
Pillow
onnxruntime
numpy
waitress
